requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.12=h4ab18f5_0
attr=2.5.1=h166bdaf_1
biopython=1.78=py312h5eee18b_0
blas=1.0=openblas
blosc=1.21.6=hef167b5_0
bottleneck=1.3.7=py312ha883a20_0
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.2=heb4867d_0
ca-certificates=2024.9.24=h06a4308_0
cairo=1.18.0=hebfffa5_3
certifi=2024.8.30=py312h06a4308_0
dbus=1.13.6=h5008d03_3
expat=2.6.3=h5888daf_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=h267a509_2
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
glew=2.1.0=h9c3ff4c_2
glib=2.82.2=h44428e9_0
glib-tools=2.82.2=h4833e2c_0
glm=0.9.9.8=h00ab1b0_0
graphite2=1.3.13=h59595ed_1003
gst-plugins-base=1.24.7=h0a52356_0
gstreamer=1.24.7=hf3bb09a_0
harfbuzz=9.0.0=hda332d3_1
hdf4=4.2.15=h2a13503_7
hdf5=1.14.4=nompi_h2d575fe_101
icu=75.1=he02047a_0
keyutils=1.6.1=h166bdaf_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
ld_impl_linux-64=2.43=h712a8e2_2
libaec=1.1.3=h59595ed_0
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libblas=3.9.0=25_linux64_openblas
libcap=2.69=h0f662aa_0
libcblas=3.9.0=25_linux64_openblas
libclang-cpp19.1=19.1.3=default_hb5137d0_0
libclang13=19.1.3=default_h9c6a7e4_0
libcups=2.3.3=h4637d8d_4
libcurl=8.10.1=hbbe4b11_0
libdrm=2.4.123=hb9d3cd8_0
libedit=3.1.20191231=he28a2e2_2
libegl=1.7.0=ha4b6fd6_1
libev=4.33=hd590300_2
libevent=2.1.12=hf998b51_1
libexpat=2.6.3=h5888daf_0
libffi=3.4.2=h7f98852_5
libflac=1.4.3=h59595ed_0
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgcrypt=1.11.0=h4ab18f5_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgl=1.7.0=ha4b6fd6_1
libglib=2.82.2=h2ff4ddf_0
libglu=9.0.0=ha6d2627_1004
libglvnd=1.7.0=ha4b6fd6_1
libglx=1.7.0=ha4b6fd6_1
libgomp=14.2.0=h77fa898_1
libgpg-error=1.50=h4f305b6_0
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=25_linux64_openblas
libllvm19=19.1.3=ha7bfdaf_0
libnetcdf=4.9.2=nompi_h2564987_115
libnghttp2=1.64.0=h161d5f1_0
libnsl=2.0.1=hd590300_0
libogg=1.3.5=h4ab18f5_0
libopenblas=0.3.28=pthreads_h94d23a6_0
libopus=1.3.1=h7f98852_1
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libpq=16.4=h2d7952a_3
libsndfile=1.2.2=hc60ed4a_1
libsqlite=3.47.0=hadc24fc_1
libssh2=1.11.0=h0841786_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libsystemd0=256.7=h2774228_1
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h9c3ff4c_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.13.4=hb346dea_2
libzip=1.11.2=h6991a6a_0
libzlib=1.3.1=hb9d3cd8_2
lz4-c=1.9.4=hcb278e6_0
mpg123=1.32.8=hc50e24c_0
mysql-common=9.0.1=h266115a_2
mysql-libs=9.0.1=he0572af_2
ncurses=6.5=he02047a_1
nspr=4.36=h5888daf_0
nss=3.106=hdf54f9c_0
numexpr=2.10.1=py312hd28fd6d_0
numpy=1.26.4=py312h2809609_0
numpy-base=1.26.4=py312he1a6c75_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=py312h526ad5a_0
pcre2=10.44=hba22ea6_2
pip=24.3.1=pyh8b19718_0
pixman=0.43.2=h59595ed_0
ply=3.11=pyhd8ed1ab_2
pmw=2.0.1=py312h7900ff3_1008
pthread-stubs=0.4=hb9d3cd8_1002
pulseaudio-client=17.0=hb77b528_0
pymol-open-source=3.0.0=py312h3c651bb_8
pyqt=5.15.9=py312h949fe66_5
pyqt5-sip=12.12.2=py312h30efb56_5
python=3.12.7=hc5c86c4_0_cpython
python-dateutil=2.9.0post0=py312h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.12=5_cp312
pytz=2024.1=py312h06a4308_0
qt-main=5.15.15=h374914d_0
readline=8.2=h8228510_1
setuptools=75.3.0=pyhd8ed1ab_0
sip=6.7.12=py312h30efb56_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=ha2e4443_0
tk=8.6.13=noxft_h4845f30_101
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.2=pyhd8ed1ab_0
tzdata=2024b=hc8b5060_0
wheel=0.44.0=pyhd8ed1ab_0
xcb-util=0.4.1=hb711507_2
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.1=hb9d3cd8_1
xorg-libsm=1.2.4=he73a12e_1
xorg-libx11=1.8.10=h4f16b4b_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxrender=0.9.11=hb9d3cd8_1
xorg-libxxf86vm=1.1.5=hb9d3cd8_4
xorg-xextproto=7.3.0=hb9d3cd8_1004
xorg-xf86vidmodeproto=2.3.1=hb9d3cd8_1004
xorg-xorgproto=2024.1=hb9d3cd8_1
xz=5.2.6=h166bdaf_0
zlib=1.3.1=hb9d3cd8_2
zstd=1.5.6=ha6fb4c9_0
